streamlit/dasheconomics/requirements.txt
numpy
pandas
seaborn
python-bcb
plotly
plotly-express
streamlit
yfinance
datetime